+++ package.json
@@ -7,6 +7,7 @@
     "@material-ui/core": "^4.6.0",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
+    "@web3-react/ledger-connector": "^6.1.9",
     "@web3-react/walletconnect-connector": "^6.2.4",
     "animated-number-react": "^0.1.1",
     "antd": "^3.26.18",
