+++ src/containers/Layout/Sidebar.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { compose } from 'recompose';
@@ -7,10 +7,8 @@ import { bindActionCreators } from 'redux';
 import { Select, Icon } from 'antd';
 import BigNumber from 'bignumber.js';
 import { useWeb3React } from '@web3-react/core';
-import * as constants from 'utilities/constants';
-import ConnectModal from 'components/Basic/ConnectModal';
+import ConnectButton from 'components/Basic/ConnectButton';
 import { Label } from 'components/Basic/Label';
-import Button from '@material-ui/core/Button';
 import { connectAccount, accountActionCreators } from 'core';
 import logoImg from 'assets/img/logo.png';
 import commaNumber from 'comma-number';
@@ -247,38 +245,6 @@ const MobileMenu = styled.div`
   }
 `;
 
-const ConnectButton = styled.div`
-  display: flex;
-  justify-content: center;
-  margin-bottom: 20px;
-
-  @media only screen and (max-width: 768px) {
-    margin: 0;
-  }
-
-  .connect-btn {
-    width: 114px;
-    height: 30px;
-    border-radius: 5px;
-    background-image: linear-gradient(to right, #f2c265, #f7b44f);
-
-    @media only screen and (max-width: 768px) {
-      width: 100px;
-    }
-
-    .MuiButton-label {
-      font-size: 13px;
-      font-weight: 500;
-      color: var(--color-text-main);
-      text-transform: capitalize;
-
-      @media only screen and (max-width: 768px) {
-        font-size: 12px;
-      }
-    }
-  }
-`;
-
 const { Option } = Select;
 const abortController = new AbortController();
 
@@ -495,21 +461,7 @@ function Sidebar({ history, settings, setSetting, getGovernanceVenus }) {
           </div>
         </TotalValue>
       )}
-      <ConnectButton>
-        <Button
-          className="connect-btn"
-          onClick={() => {
-            setIsOpenModal(true);
-          }}
-        >
-          {!account
-            ? 'Connect'
-            : `${account.substr(0, 6)}...${account.substr(
-                account.length - 4,
-                4
-              )}`}
-        </Button>
-      </ConnectButton>
+      <ConnectButton />
       <MobileMenu id="main-menu">
         <Select
           defaultValue={defaultPath}
@@ -560,10 +512,6 @@ function Sidebar({ history, settings, setSetting, getGovernanceVenus }) {
           )}
         </Select>
       </MobileMenu>
-      <ConnectModal
-        visible={isOpenModal}
-        onCancel={() => setIsOpenModal(false)}
-      />
     </SidebarWrapper>
   );
 }
